updates to how start_type = new is handled

+++ forcing/ocn00/make_forcing_main.py
@@ -351,14 +351,15 @@ ds.close()
 print('- Write clm file: %0.2f sec' % (time()-tt0))
 sys.stdout.flush()
 
-# Write initial condition file
-tt0 = time()
-in_fn = out_dir / 'ocean_clm.nc'
-out_fn = out_dir / 'ocean_ini.nc'
-out_fn.unlink(missing_ok=True)
-Ofun_nc.make_ini_file(in_fn, out_fn)
-print('- Write ini file: %0.2f sec' % (time()-tt0))
-sys.stdout.flush()
+if Ldir['start_type'] == 'new':
+    # Write initial condition file if needed
+    tt0 = time()
+    in_fn = out_dir / 'ocean_clm.nc'
+    out_fn = out_dir / 'ocean_ini.nc'
+    out_fn.unlink(missing_ok=True)
+    Ofun_nc.make_ini_file(in_fn, out_fn)
+    print('- Write ini file: %0.2f sec' % (time()-tt0))
+    sys.stdout.flush()
 
 # Write boundary file
 tt0 = time()
@@ -376,13 +377,18 @@ def print_info(fn):
     ds.close()
 
 # Check results
-nc_list = ['ocean_clm.nc', 'ocean_ini.nc', 'ocean_bry.nc']
+if Ldir['start_type'] == 'new':
+    nc_list = ['ocean_clm.nc', 'ocean_ini.nc', 'ocean_bry.nc']
+else:
+    nc_list = ['ocean_clm.nc', 'ocean_bry.nc']
 if Ldir['testing']:
     # print info about the files to the screen
     # for fn in nc_list:
     #     print_info(out_dir / fn)
+    # open datasets to have a peek
     dsc = xr.open_dataset(out_dir / 'ocean_clm.nc', decode_times=False)
-    dsi = xr.open_dataset(out_dir / 'ocean_ini.nc', decode_times=False)
+    if Ldir['start_type'] == 'new':
+        dsi = xr.open_dataset(out_dir / 'ocean_ini.nc', decode_times=False)
     dsb = xr.open_dataset(out_dir / 'ocean_bry.nc', decode_times=False)
         
 result_dict['result'] = 'success'
